+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 ## Changelog
 
+### Version 1.43
+
+- Fix for NX projects when node modules are not installed
+- Fix for the build command when an Angular configuration is selected
+- Fix for Angular project with multiple configurations
+
 ### Version 1.42
 
 - Add recommendation for cordova-sqlite-storage causing errors with Secure Storage